+++ include/lbann/execution_algorithms/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Add the headers for this directory
 set_full_path(THIS_DIR_HEADERS
-  batch_inference_algorithm.hpp
+  batch_functional_inference_algorithm.hpp
   training_algorithm.hpp
   sgd_training_algorithm.hpp
   )
