+++ include/lbann/execution_algorithms/batch_functional_inference_algorithm.hpp
