+++ src/execution_algorithms/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Add the source files for this directory
 set_full_path(THIS_DIR_SOURCES
-  batch_inference_algorithm.cpp
+  batch_functional_inference_algorithm.cpp
   sgd_training_algorithm.cpp
   training_algorithm.cpp
   )
